requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
opencv-python
albumentations
scikit-learn
monai
tqdm
einops
scipy
pandas
ipykernel
streamlit
pillow
